requirement.txt
aiofiles==24.1.0
aiosqlite==0.20.0
alembic==1.14.1
altgraph==0.17.3
annotated-types==0.7.0
anyio==4.0.0
appdirs==1.4.4
APScheduler==3.11.0
argon2-cffi==20.1.0
asgiref==3.8.1
async-timeout==4.0.2
asyncqt==0.8.0
attrs==25.3.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.61
bcrypt==3.2.0
bleach==4.0.0
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
channels==4.2.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
cx-Oracle==8.3.0
cycler==0.10.0
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.2
Django==4.2.21
django-channels==0.7.0
django-import-export==4.1.1
django-jazzmin==3.0.1
dnspython==2.4.2
docx==0.2.4
docxcompose==1.4.0
docxtpl==0.19.1
download==0.3.5
ecdsa==0.19.1
email-validator==2.1.0.post1
emoji==2.12.1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.97.0
filelock==3.0.12
flake8==3.9.2
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Babel==2.0.0
Flask-SQLAlchemy==3.0.3
future==0.18.2
gpt4all==2.0.2
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
identify==2.2.10
idna==2.10
imageio==2.9.0
importlib-metadata==6.0.0
importlib_resources==6.4.5
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.4.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
key==0.4
kiwisolver==1.3.1
loguru==0.7.3
lxml==4.6.3
Mako==1.3.9
Markdown==3.7
MarkupSafe==2.1.2
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
MouseInfo==0.1.3
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nodeenv==1.6.0
notebook==6.4.3
numpy==1.21.0
oauthlib==3.2.2
opencv-python==4.5.2.54
openpyxl==3.1.2
orjson==3.9.10
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
passlib==1.7.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.4.0
pre-commit==2.13.0
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2-binary==2.9.9
pyasn1==0.4.8
PyAutoGUI==0.9.54
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.10.9
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyecharts==1.9.0
PyExecJS==1.5.1
pyflakes==2.3.1
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.9.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.4.0
pypandoc==1.15
pyparsing==2.4.7
pyperclip==1.8.2
pypng==0.20220715.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.2.0
pyrsistent==0.18.0
PyScreeze==0.1.30
python-dateutil==2.8.1
python-docx==1.1.2
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
pytweening==1.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.3
PyX==0.16
PyYAML==5.4.1
pyzmq==22.2.1
qrcode==7.4.2
qtconsole==5.1.1
QtPy==1.10.0
redis==4.5.5
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
scapy==2.5.0
scikit-image==0.17.2
scikit-learn==1.3.2
scipy==1.7.0
selenium==3.141.0
Send2Trash==1.8.0
Shapely==1.7.1
shellcheck-py==0.7.2.1
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
some-package==0.1
SQLAlchemy==2.0.3
sqlparse==0.5.3
starlette==0.27.0
tablib==3.5.0
tenacity==8.5.0
terminado==0.11.1
testpath==0.5.0
thekey==2.0.2
threadpoolctl==3.5.0
tifffile==2021.6.14
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing_extensions==4.12.2
tzdata==2022.6
tzlocal==4.2
ujson==5.9.0
urllib3==1.26.5
uvicorn==0.33.0
vboxapi==1.0
virtualenv==20.4.7
visualdl==2.2.0
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==12.0
Werkzeug==2.2.2
widgetsnbextension==3.5.1
win32_setctime==1.2.0
zipp==3.12.1
